fix(rust, python): fix logical columns of streaming multi-column sort
